requirements.txt
asyncio==3.4.3 ; python_version >= "3.9" and python_version < "4.0"
babel==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.11.2 ; python_version >= "3.9" and python_version < "4.0"
bracex==2.3.post1 ; python_version >= "3.9" and python_version < "4.0"
brotli==1.0.9 ; platform_python_implementation == "CPython" and python_version >= "3.9" and python_version < "4.0"
brotlicffi==1.0.9.2 ; platform_python_implementation != "CPython" and python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "4.0"
chardet==3.0.4 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "4"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
csscompressor==0.9.5 ; python_version >= "3.9" and python_version < "4.0"
cssselect2==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
cssselect==1.2.0 ; python_version >= "3.9" and python_version < "4.0"
docxcompose==1.4.0 ; python_version >= "3.9" and python_version < "4.0"
editorconfig==0.12.3 ; python_version >= "3.9" and python_version < "4.0"
fonttools[woff]==4.38.0 ; python_version >= "3.9" and python_version < "4.0"
ghp-import==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
gitdb==4.0.10 ; python_version >= "3.9" and python_version < "4.0"
gitpython==3.1.30 ; python_version >= "3.9" and python_version < "4.0"
googletrans==3.1.0a0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.9.0 ; python_version >= "3.9" and python_version < "4.0"
h2==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
hpack==3.0.0 ; python_version >= "3.9" and python_version < "4.0"
hstspreload==2022.12.1 ; python_version >= "3.9" and python_version < "4.0"
html5lib==1.1 ; python_version >= "3.9" and python_version < "4.0"
htmlmin==0.1.12 ; python_version >= "3.9" and python_version < "4.0"
httpcore==0.9.1 ; python_version >= "3.9" and python_version < "4.0"
httpx==0.13.3 ; python_version >= "3.9" and python_version < "4.0"
hyperframe==5.2.0 ; python_version >= "3.9" and python_version < "4.0"
idna==2.10 ; python_version >= "3.9" and python_version < "4"
importlib-metadata==5.1.0 ; python_version >= "3.9" and python_version < "3.10"
jieba==0.42.1 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
jsbeautifier==1.14.7 ; python_version >= "3.9" and python_version < "4.0"
jsmin==3.0.1 ; python_version >= "3.9" and python_version < "4.0"
libsass==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
lxml==4.9.1 ; python_version >= "3.9" and python_version < "4.0"
markdown2==2.4.6 ; python_version >= "3.9" and python_version < "4"
markdown==3.3.7 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
mergedeep==1.3.4 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-awesome-pages-plugin==2.8.0 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-embed-external-markdown==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-git-authors-plugin==0.6.5 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-git-committers-plugin-2==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-git-revision-date-localized-plugin==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-git-revision-date-plugin==0.3.2 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-macros-plugin==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-material @ git+https://github.com/daocloud/mkdocs-material-insiders.git@HEAD ; python_version >= "3.9" and python_version < "4.0"
mkdocs-material-extensions==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-mermaid2-plugin==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-minify-plugin==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-multirepo-plugin==0.4.12 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-print-site-plugin==2.3.4 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-redirects==1.2.0 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-rss-plugin==1.6.0 ; python_version >= "3.9" and python_version < "4"
mkdocs-swagger-ui-tag==0.6.1 ; python_version >= "3.9" and python_version < "4.0"
mkdocs-with-pdf==0.9.3 ; python_version >= "3.9" and python_version < "4.0"
mkdocs==1.4.2 ; python_version >= "3.9" and python_version < "4.0"
natsort==8.2.0 ; python_version >= "3.9" and python_version < "4.0"
packaging==22.0 ; python_version >= "3.9" and python_version < "4.0"
paginate==0.5.6 ; python_version >= "3.9" and python_version < "4.0"
pandoc==2.3 ; python_version >= "3.9" and python_version < "4.0"
pillow==9.3.0 ; python_version >= "3.9" and python_version < "4.0"
plumbum==1.8.1 ; python_version >= "3.9" and python_version < "4.0"
ply==3.11 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pydyf==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
pygments==2.13.0 ; python_version >= "3.9" and python_version < "4.0"
pymdown-extensions==9.9 ; python_version >= "3.9" and python_version < "4.0"
pyphen==0.13.2 ; python_version >= "3.9" and python_version < "4.0"
pyquery==1.4.3 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-docx==0.8.11 ; python_version >= "3.9" and python_version < "4.0"
python-slugify==7.0.0 ; python_version >= "3.9" and python_version < "4.0"
pytz==2022.6 ; python_version >= "3.9" and python_version < "4.0"
pywin32==305 ; platform_system == "Windows" and platform_python_implementation != "PyPy" and python_version >= "3.9" and python_version < "4.0"
pyyaml-env-tag==0.1 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
readtime==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
regex==2022.10.31 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.9" and python_version < "4"
rfc3986==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
setuptools==65.6.3 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
smmap==5.0.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.3.2.post1 ; python_version >= "3.9" and python_version < "4.0"
termcolor==2.2.0 ; python_version >= "3.9" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.9" and python_version < "4.0"
tinycss2==1.2.1 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.9" and sys_platform == "win32" and python_version < "4"
ufile==3.2.6 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.13 ; python_version >= "3.9" and python_version < "4"
watchdog==2.2.0 ; python_version >= "3.9" and python_version < "4.0"
wcmatch==8.4.1 ; python_version >= "3.9" and python_version < "4.0"
weasyprint==57.1 ; python_version >= "3.9" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.11.0 ; python_version >= "3.9" and python_version < "3.10"
zopfli==0.2.2 ; python_version >= "3.9" and python_version < "4.0"
